requirements.txt
# Web framework
flask
gunicorn
python-dotenv

# Auth
authlib
requests

# Optional: pywebview only needed for the desktop build, not the web server
# pywebview
